Fixed navigation with path bar.
TextChanged was firing as the user types, and it seemed to be firing before the model was updated, resulting in it not going anywhere. Now the TextChanged is ignored while the box is in focus so that we send a PathChanged only when it is changed from the model. A PathChanged is also sent when the path box loses focus to handle updating after the user is done editing the path.