Changed Path bind to one way, send value thru evt.
Instead of relying on a weird feedback loop to open paths from the controller, it is now explicit.

Before, the controller would set the model Path, which would update the text box, which would fire a PathChanged event, which would normalize the path and fire the event again, then load the nodes. This was difficult to reason about and had odd effects, such as the path box changed events only triggering after the current event was completed processing.

The path binding now only goes from the model to the text box, and the only trigger for OpenPath event is pressing Enter in the text box. Other event handlers open a path by calling the controller's OpenPath function.