Major refactor of path handling.
Before, Path was a single-case union wrapper with no restrictions on instantiation, with all operations on paths defined in FileSystemService. The values were dependent on the current path format, and changing the path format meant re-interpreting the paths and reloading nodes.

Now, Path is a class that controls what values are valid can only be instantiated through Parse, which returns a Path option. It has one internal format (Windows) and defines all the operations on path values. Values are re-formatted only for display.