Changes projects to error on unused expr. value.
When an expression in a function that is not last has a type other than unit, you get the warning "FS0020: This expression should have type 'unit', but has type X". Since this is usually unintentional, an explicit pipe into `ignore` should be used when it is intentional. I am changing the projects to treat these as errors so that the programmer is forced to ignore or capture the expression value explicitly.