Moves OpenPath into MainHandler module.
This is the first of many changes to move code under test into plain module functions with simple function dependencies. This will make the tests simpler and the controller will just dispatch to the handler module functions, providing the function dependencies. This will also avoid the use of mocks, which have turned out to be a real pain.