Renames parameters to undo functions for accuracy.
The parameters were named according to what the values were at the time the action was recorded (node, newPath) instead of what the values meant when the action was being undone (prevNode, currentPath). This was a bit confusing and is probably the cause of the bug I just noticed and added a TODO for.