Changes scoop manifest to be a package artifact.
I am removing the scoop manifest from the root directory, trusting that the official extras bucket will be kept up-to-date automatically. Instead, the package step will output it to the dist folder with the other package artifacts where it can be used to re-submit to the extras bucket repo if something other than the version number changes. Also, it can be made part of the github release artifacts if deemed necessary.