Removes TODO to change Nodes to array.
I changed Nodes to an array, simply changing List functions to Array functions where it was passed in. I
benchmarked this against the old list version by loading a folder with 1700 items and searching it and saw no difference.

Maybe there would be optimizations to make where an operation could use array-specific functions, but I'm not sure where and it seems responsive enough as it is.