Fixes refresh opening root when in an empty folder.
I changed refresh to open the current item's parent when i added sub-folder searching, but looking at it now, this doesn't make any sense in any context. The refresh command actually performs the search again when there is an active search instead of calling the refresh function, and the other calls of refresh all assume that it will open the current path. The only time it makes a difference is when the selected item is of the Empty type, in which case it's wrong to use its parent path because it's always root.