Revert "Swaps key bindings for create file, create folder"
This reverts commit bb16a98cd7a123dd984ed08a6865ca363e5e5afb.

Swapping the bindings seemed to make sense because creating folders is more frequently used and should be the easier to use `o` binding. However even after several months I still try to use the original bindings. It still feels more natural, I guess because in Vim, `o` creates a new line below and `O` creates a line above, which correspond to the default ordering of files and folders respectively.