Reorders bindings to avoid status update cancelling input
The binding for status focuses the grid when coming back from a Busy state, but since the input mode binding came before that and could open and focus the input, this meant the input lost focus triggering a CancelInput event.

Simply moving the status binding before the input mode binding fixes this unexpected conflict.