Fixes window focus event firing when focus was lost instead
The "Prevent loss of keyboard focus" code was firing when window focus was lost, immediately triggering the window Activated event. When the window was actually activated, the event would not fire again.

This commit replaces two separate focus fixes with a simpler one that focuses the grid on window focus if none of the focusable controls are focused. It also replaces a Settings button focus fix by making the settings button not focusable.